refactor(client): unify Client — clientV2 becomes the only Client
- Replaced client.ts with clientV2's Dispatcher-based implementation
- Ported: applyElicitationDefaults + setRequestHandler validation, listChanged debounce,
  getSupportedElicitationModes, ClientTasksCapabilityWithRuntime, ClientOptions extends ProtocolOptions
- pipeAsClientTransport accepts StreamDriverOptions (tasks config flows through)
- Deleted clientV2.ts, renamed test file
- Fixed onclose double-fire, task option threading, cancel SdkError type

All 1537 tests + lint + typecheck green.
